% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\docType{data}
\name{hash_lemma_de}
\alias{hash_lemma_de}
\title{German Lemma Dictionary}
\format{A data frame with 358473 rows and 2 variables:
\describe{
  \item{word}{a key word}
  \item{lemma}{its mapped lemma.}
  ...
}}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
hash_lemma_de
}
\description{
German Lemma Dictionary
}
\keyword{datasets}
