% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\docType{data}
\name{sent_afinn}
\alias{sent_afinn}
\title{AFINN Sentiment Dictionary}
\format{A data frame with 2,476 rows and 2 variables:
\describe{
  \item{word}{a key word}
  \item{afinn_score}{ sentiment number }
  ...
}}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
sent_afinn
}
\description{
AFINN Sentiment Dictionary
}
\keyword{datasets}
