% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\docType{data}
\name{sent_vader}
\alias{sent_vader}
\title{Vader Sentiment Dictionary}
\format{A data frame with 7,516 rows and 6 variables:
\describe{
  \item{word}{a key word}
  \item{vader_score}{ sentiment number }
  \item{vader_intens}{ sentiment number }
  \item{vader_seq}{ sentiment number }
  \item{vader_med}{ sentiment number }
  \item{vader_class}{ sentiment class }
  ...
}}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
sent_vader
}
\description{
Vader Sentiment Dictionary
}
\keyword{datasets}
