% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\docType{data}
\name{sent_wordnet}
\alias{sent_wordnet}
\title{Wordnet Sentiment Dictionary}
\format{A data frame with 24,428 rows and 3 variables:
\describe{
  \item{word}{a key word}
  \item{wordnet_score}{ sentiment number }
  \item{wordnet_class}{ sentiment class }
  ...
}}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
sent_wordnet
}
\description{
Wordnet Sentiment Dictionary
}
\keyword{datasets}
