% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\docType{data}
\name{stop_words_en}
\alias{stop_words_en}
\title{English stopwords}
\format{A data frame with 358473 rows and 2 variables:
\describe{
  \item{word}{a key word}
  \item{lemma}{its mapped lemma.}
  ...
}}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
stop_words_en
}
\description{
English stopwords
}
\keyword{datasets}
