% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_word2vec.R
\name{tx_get_dtm}
\alias{tx_get_dtm}
\title{tx_get_dtm}
\usage{
tx_get_dtm(text, id, vec)
}
\arguments{
\item{text}{A string vector}

\item{id}{a unique identifier}

\item{vec}{a pre built word vectorizer}
}
\value{
a dtm
}
\description{
tx_get_dtm
}
