% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_keras.R
\name{tx_keras_predict}
\alias{tx_keras_predict}
\title{tx_keras_predict}
\usage{
tx_keras_predict(model, seq, index = 1)
}
\arguments{
\item{model}{a pretrained keras model}

\item{index}{if index + 1?}

\item{data}{a list of data objects}
}
\value{
a target vector
}
\description{
tx_keras_predict
}
