% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_keras.R
\name{tx_keras_predict_one}
\alias{tx_keras_predict_one}
\title{tx_keras_predict_one}
\usage{
tx_keras_predict_one(model, string, token_fun, maxlen)
}
\arguments{
\item{model}{a pre-trained model}

\item{string}{a one element string}

\item{token_fun}{a pre-constructed tokenizer}
}
\value{
a \code{x}
}
\description{
tx_keras_predict_one
}
