% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_keras.R
\name{tx_keras_predict_str}
\alias{tx_keras_predict_str}
\title{tx_keras_predict_str}
\usage{
tx_keras_predict_str(string, str, maxlen)
}
\arguments{
\item{string}{a one element string}

\item{str}{a list structure of parts brought by different models}
}
\value{
data frame of predictions \code{x}
}
\description{
tx_keras_predict_str
}
