% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_n_tokens}
\alias{tx_n_tokens}
\title{tx_n_tokens}
\usage{
tx_n_tokens(string, token = "word")
}
\arguments{
\item{x}{A string vector}
}
\value{
counts tokens of string\code{x}
}
\description{
tx_n_tokens
}
\examples{
"Oh my god? #RealDonaldTrump is a moron!" \%>\%
   tx_n_tokens()
}
