% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_keras.R
\name{tx_onehot}
\alias{tx_onehot}
\title{tx_k_onehot}
\usage{
tx_onehot(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
transforms a k-level factor vector into a n*k matrix (one-hote encoding)\code{x}
}
\description{
tx_k_onehot
}
