% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_party_de}
\alias{tx_party_de}
\title{tx_party_de}
\usage{
tx_party_de(x, ...)
}
\arguments{
\item{x}{a ggplot2 plot}
}
\value{
a color palette\code{x}
}
\description{
tx_party_de
}
\examples{
"Oh my god? #RealDonaldTrump is a moron!" \%>\%
   tx_n_tokens()
}
