% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_replace_currency}
\alias{tx_replace_currency}
\title{tx_replace_currency}
\usage{
tx_replace_currency(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
currency symbols replaced in \code{x}
}
\description{
tx_replace_currency
}
\examples{
"Oh my god? the product costs ¥30, £11, €10, $5 ..." \%>\%
   tx_replace_currency()
}
