% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_replace_emoji}
\alias{tx_replace_emoji}
\title{tx_replace_emoji}
\usage{
tx_replace_emoji(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
emoji corrected \code{x}
}
\description{
tx_replace_emoji
}
\examples{
"Hi :D my god. My brain hurts :/" \%>\%
   tx_replace_emoji()
}
