% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_replace_str}
\alias{tx_replace_str}
\title{tx_replace_str}
\usage{
tx_replace_str(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
`\n` `\t` replaced as well as to_lower in \code{x}
}
\description{
tx_replace_str
}
\examples{
"Oh my god \\n the product costs \\t ..." \%>\%
   tx_replace_str()
}
