% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_replace_twitter}
\alias{tx_replace_twitter}
\title{tx_replace_twitter}
\usage{
tx_replace_twitter(x, rm_hash = T, replace_hash = F, replace_hndl = T,
  rm_retweet = T)
}
\arguments{
\item{x}{A string vector}
}
\value{
emoji corrected \code{x}
}
\description{
tx_replace_twitter
}
\examples{
"RT @meee: Oh my god. #RealDonaldTrump is a moron :D" \%>\%
   tx_replace_twitter()
}
