% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_replace_url}
\alias{tx_replace_url}
\title{tx_replace_url}
\usage{
tx_replace_url(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
cleaned \code{x}
}
\description{
tx_replace_url
}
\examples{
"Hi http://google.com my god" \%>\%
   tx_replace_url()
}
