% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_sentiment.R
\name{tx_sent}
\alias{tx_sent}
\title{tx_sent}
\usage{
tx_sent(data, string, lexica = c("afinn"))
}
\arguments{
\item{x}{A string vector}
}
\value{
seq of vectorized character vectors as list\code{x}
}
\description{
tx_sent
}
