% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_replace.R
\name{tx_spacing}
\alias{tx_spacing}
\title{tx_spacing}
\usage{
tx_spacing(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
optimized white space in \code{x}
}
\description{
tx_spacing
}
\examples{
"  Oh my god?   #RealDonaldTrump    is a moron!" \%>\%
   tx_spacing()
}
