% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_keras.R
\name{tx_text_to_seq}
\alias{tx_text_to_seq}
\title{tx_keras_text_seq}
\usage{
tx_text_to_seq(token_fun, string, maxlen)
}
\arguments{
\item{x}{A string vector}
}
\value{
seq of vectorized character vectors as list\code{x}
}
\description{
tx_keras_text_seq
}
