% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_map.R
\name{tx_map_dict}
\alias{tx_map_dict}
\title{tx_map_dict}
\usage{
tx_map_dict(string, dict, iter_by = "for", key1 = 1, key2 = 2)
}
\arguments{
\item{x}{A string vector}
}
\value{
mapped dictionary on string for replacement\code{x}
}
\description{
tx_map_dict
}
